Fix null message regression after quiet logger addition (#44)
The `Microsoft.Build.Utilities.Core` LogError requires a non-null
string. In the refactoring, the usages of LogError fell into a new
overload of LogError which was passing a null message as the first two
arguments of LogError signature were the same.

This removes the old logger and adds more protections for this case.